moved project files to TNN folder
